layout/generic/test/test_bug666225.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=666225
-->
<head>
  <title>Test for Bug 666225</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script> 
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=666225">Mozilla Bug 666225</a>
<p id="display"><input id="i" onmousedown="this.type='file';"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">
/** Test for Bug 666225
    Causes ASSERTION: Unexpected document: 'capturingContent->GetUncomposedDoc() == GetDocument()' (bug 560764) **/

function click() 
{
    $("i").addEventListener("mouseup", test, false);
    synthesizeMouseAtCenter($("i"), { type: "mousedown" });
    synthesizeMouseAtCenter($("i"), { type: "mouseup" });
}

function test() 
{
    ok(true, "should not crash");
    SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(click);

</script>
</pre>
</body>
</html>
